+++ Plugins/Creature/Creature.cpp
@@ -3573,3 +3573,12 @@ NWNX_EXPORT ArgumentStack SetAbilityIncreaseByLevel(ArgumentStack&& args)
 
     return {};
 }
+
+NWNX_EXPORT ArgumentStack GetMaxAttackRange(ArgumentStack&& args)
+{
+    auto* pCreature = Utils::PopCreature(args);
+    ASSERT_OR_THROW(pCreature);
+    auto oidTarget = args.extract<OBJECT_ID>();
+
+    return pCreature->MaxAttackRange(oidTarget);
+}